validated the correctness of the algorithm; converted the algorithm to float to prove correctness with monte carlos (i.e. random samples).

+++ lowbit_matmul/benchmark/check_correctness.py
@@ -43,7 +43,8 @@ def pack_i4(q):
 
 # %%
 
-A = torch.arange(size, dtype=torch.half).reshape(-1, size)
+# A = torch.arange(size, dtype=torch.half).reshape(-1, size)
+A = torch.arange(size, dtype=torch.float32).reshape(-1, size)
 B = torch.eye(size, dtype=torch.int8).cuda()
 
 y = lowbit_matmul.kernels.matmul_hi4(A.cuda(), pack_i4(B).cuda())
